+++ src/Cpu/Interrupts.h
@@ -22,7 +22,7 @@ static const uint16_t g_interrupt_request_address = 0xFF0F;
 static const uint16_t g_interrupt_enabled_address = 0xFFFF;
 
 int handle_interrupts(Cpu* cpu);
-void call_interrupt_vector(Cpu* cpu, uint8_t interrupt_vector);  
+void call_interrupt_vector(Cpu* cpu, uint16_t interrupt_vector);  
 uint16_t to_vector(Interrupts interrupt);
 bool is_enabled(MemoryBankController* mc, Interrupts interrupt);
 bool is_requested(MemoryBankController* mc, Interrupts interrupt);
